// src/types/memberComplete.ts
export interface CompleteProfileData {

    birthDate: string

    phone: string

    address: string

    nif: string

    profession: string

    institution: string

    district: string

    professionalLicenseNumber?: string

    specialties: string[]

    habilitacoes: string[]

    profilePhotoUrl?: string

    professionalCardUrl?: string

    certificatesUrls?: string[]
}